+++ NEWS.md
@@ -1,3 +1,7 @@
+# gibble dev
+
+* Fixes for SpatialLines and SpatialMultiPoints, found in anglr. 
+
 # gibble 0.0.2
 
 * Added `PATH`, `sp` , `trip` methods for `gibble`.